Endpoint for Editting the post works now, the frontend needs updating, the postscreen needs updates too since it uses a lot of the same logic
